update handling of string literals to work with single and double quotes in line with other routines
isolate functional code sets
add unit tests for string literal handling
