+++ src/helper.js
@@ -1,0 +1,9 @@
+function inHex(n, l) {
+    var s = n.toString(16).toUpperCase();
+    while (s.length < l) {
+        s = '0' + s;
+    }
+    return s;
+}
+
+export {inHex}
